third_party/blink/web_tests/platform/generic/external/wpt/html/webappapis/scripting/processing-model-2/runtime-error-same-origin-with-hash-expected.txt

This is a testharness.js-based test.
FAIL window.onerror - runtime error in <script src=...> with hash assert_equals: second arg expected "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/support/undefined-variable.js#" but got "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/support/undefined-variable.js"
PASS window.onerror - runtime error in <script src=...> with hash (column)
Harness: the test ran to completion.
